+++ src/main/java/org/mmarini/wheelly/engines/Builders.java
@@ -34,17 +34,21 @@ import org.mmarini.wheelly.engines.statemachine.*;
 import org.mmarini.wheelly.model.InferenceEngine;
 import org.mmarini.wheelly.swing.RxJoystick;
 import org.mmarini.wheelly.swing.RxJoystickImpl;
+import org.mmarini.wheelly.swing.Yaml;
+import org.mmarini.yaml.schema.Locator;
 
 import java.awt.geom.Point2D;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import static org.mmarini.Utils.stream;
 import static org.mmarini.wheelly.engines.statemachine.ContextOperator.assign;
 import static org.mmarini.wheelly.engines.statemachine.ContextOperator.remove;
 import static org.mmarini.wheelly.engines.statemachine.EngineStatus.*;
+import static org.mmarini.wheelly.engines.statemachine.FindPathStatus.NO_PATH_EXIT;
+import static org.mmarini.wheelly.engines.statemachine.FindPathStatus.TARGET_REACHED_EXIT;
 import static org.mmarini.wheelly.engines.statemachine.GotoStatus.UNREACHABLE_EXIT;
 import static org.mmarini.wheelly.engines.statemachine.StateMachineContext.TIMEOUT_KEY;
+import static org.mmarini.wheelly.swing.Yaml.point;
+import static org.mmarini.wheelly.swing.Yaml.points;
 
 public interface Builders {
 
@@ -64,24 +68,8 @@ public interface Builders {
     }
 
     static InferenceEngine findPath(JsonNode config) {
-        JsonNode targetsNode = config.path("targets");
-        if (targetsNode.isMissingNode()) {
-            throw new IllegalArgumentException("Missing targets");
-        }
-        if (!targetsNode.isArray()) {
-            throw new IllegalArgumentException("targets must be an array");
-        }
-        List<Point2D> targets = stream(targetsNode.elements()).map(node -> {
-            JsonNode xNode = node.path("x");
-            JsonNode yNode = node.path("y");
-            if (xNode.isMissingNode()) {
-                throw new IllegalArgumentException("Missing x");
-            }
-            if (yNode.isMissingNode()) {
-                throw new IllegalArgumentException("Missing y");
-            }
-            return new Point2D.Double(xNode.asDouble(), yNode.asDouble());
-        }).collect(Collectors.toList());
+        point().apply(Locator.root().path("target")).accept(config);
+        Point2D target = point(config.path("target")).orElseThrow();
         return StateMachineBuilder.create()
                 .addState("initial", StopStatus.create())
                 .addState("scan", ScanStatus.create())
@@ -95,46 +83,36 @@ public interface Builders {
                         ContextOperator.assignValue(TIMEOUT_KEY, 10000))
                 .addTransition("scan", COMPLETED_EXIT, "findPath")
                 .addTransition("scan", OBSTACLE_EXIT, "avoid")
-                .addTransition("findPath", FindPathStatus.PATH_EXIT, "next",
+                .addTransition("findPath", COMPLETED_EXIT, "next",
                         ContextOperator.sequence(
                                 assign(NextSequenceStatus.LIST_KEY, FindPathStatus.PATH_KEY),
                                 remove(NextSequenceStatus.INDEX_KEY)))
                 .addTransition("next", NextSequenceStatus.TARGET_SELECTED_EXIT, "goto",
                         assign(GotoStatus.TARGET_KEY, NextSequenceStatus.TARGET_KEY))
-                .addTransition("goto", GotoStatus.TARGET_REACHED_EXIT, "next")
+                .addTransition("goto", COMPLETED_EXIT, "next")
                 .addTransition("goto", OBSTACLE_EXIT, "avoid")
                 .addTransition("goto", UNREACHABLE_EXIT, "findPath")
                 .addTransition("avoid", COMPLETED_EXIT, "safe")
                 .addTransition("safe", COMPLETED_EXIT, "gotoSafe",
                         assign(GotoStatus.TARGET_KEY, NearestSafeStatus.TARGET_KEY))
-                .addTransition("gotoSafe", GotoStatus.TARGET_REACHED_EXIT, "scan")
+                .addTransition("gotoSafe", COMPLETED_EXIT, "scan")
                 .addTransition("gotoSafe", OBSTACLE_EXIT, "avoid")
                 .addTransition("gotoSafe", UNREACHABLE_EXIT, "safe")
-                .setParams(FindPathStatus.TARGET_KEY, targets.get(0))
+                .setParams(FindPathStatus.TARGET_KEY, target)
                 .setParams(TIMEOUT_KEY, 2000)
                 .build("initial");
     }
 
     static InferenceEngine gotoTest(JsonNode config) {
-        JsonNode targetsNode = config.path("targets");
-        if (targetsNode.isMissingNode()) {
-            throw new IllegalArgumentException("Missing targets");
-        }
-        if (!targetsNode.isArray()) {
-            throw new IllegalArgumentException("targets must be an array");
-        }
-        List<Point2D> targets = stream(targetsNode.elements()).map(node -> {
-            JsonNode xNode = node.path("x");
-            JsonNode yNode = node.path("y");
-            if (xNode.isMissingNode()) {
-                throw new IllegalArgumentException("Missing x");
-            }
-            if (yNode.isMissingNode()) {
-                throw new IllegalArgumentException("Missing y");
-            }
-            return new Point2D.Double(xNode.asDouble(), yNode.asDouble());
-        }).collect(Collectors.toList());
-        return StateMachineBuilder.create().addState("initial", StopStatus.create()).addState("goto", GotoStatus.create()).addTransition("initial", TIMEOUT_EXIT, "goto").setParams(GotoStatus.TARGET_KEY, targets.get(0)).setParams(TIMEOUT_KEY, 4000).build("initial");
+        point().apply(Locator.root().path("target")).accept(config);
+        Point2D target = point(config.path("target")).orElseThrow();
+        return StateMachineBuilder.create()
+                .addState("initial", StopStatus.create())
+                .addState("goto", GotoStatus.create())
+                .addTransition("initial", TIMEOUT_EXIT, "goto")
+                .setParams(GotoStatus.TARGET_KEY, target)
+                .setParams(TIMEOUT_KEY, 4000)
+                .build("initial");
     }
 
     static InferenceEngine manual(JsonNode config) {
@@ -146,25 +124,81 @@ public interface Builders {
         return new ManualEngine(joystick);
     }
 
+    static InferenceEngine randomPath(JsonNode config) {
+        Yaml.randomPath().apply(Locator.root()).accept(config);
+
+        StateMachineBuilder builder1 = StateMachineBuilder.create()
+                .addState("initial", StopStatus.create())
+                .addState("scanForTarget", ScanStatus.create())
+                .addState("avoid1", AvoidObstacleStatus.create())
+                .addState("safe1", NearestSafeStatus.create())
+                .addState("gotoSafe1", GotoStatus.create())
+                .addState("random", RandomTargetStatus.create())
+                .addState("findPath", FindPathStatus.create())
+                .addState("next", NextSequenceStatus.create())
+                .addState("goto", GotoStatus.create())
+                .addState("avoid2", AvoidObstacleStatus.create())
+                .addState("safe2", NearestSafeStatus.create())
+                .addState("gotoSafe2", GotoStatus.create())
+                .addState("scanForPath", ScanStatus.create())
+
+                .addTransition("initial", TIMEOUT_EXIT, "scanForTarget",
+                        ContextOperator.assignValue(TIMEOUT_KEY, 10000))
+
+                .addTransition("scanForTarget", COMPLETED_EXIT, "random")
+                .addTransition("scanForTarget", OBSTACLE_EXIT, "avoid1")
+
+                .addTransition("avoid1", COMPLETED_EXIT, "safe1")
+                .addTransition("safe1", COMPLETED_EXIT, "gotoSafe1",
+                        assign(GotoStatus.TARGET_KEY, NearestSafeStatus.TARGET_KEY))
+                .addTransition("gotoSafe1", COMPLETED_EXIT, "scanForPath")
+                .addTransition("gotoSafe1", OBSTACLE_EXIT, "avoid1")
+                .addTransition("gotoSafe1", UNREACHABLE_EXIT, "safe1")
+
+                .addTransition("random", COMPLETED_EXIT, "findPath",
+                        assign(FindPathStatus.TARGET_KEY, RandomTargetStatus.TARGET_KEY))
+
+                .addTransition("findPath", COMPLETED_EXIT, "next",
+                        ContextOperator.sequence(
+                                assign(NextSequenceStatus.LIST_KEY, FindPathStatus.PATH_KEY),
+                                remove(NextSequenceStatus.INDEX_KEY)))
+                .addTransition("findPath", NO_PATH_EXIT, "random")
+                .addTransition("findPath", TARGET_REACHED_EXIT, "random")
+
+                .addTransition("next", NextSequenceStatus.TARGET_SELECTED_EXIT, "goto",
+                        assign(GotoStatus.TARGET_KEY, NextSequenceStatus.TARGET_KEY))
+                .addTransition("next", COMPLETED_EXIT, "scanForTarget")
+
+                .addTransition("goto", COMPLETED_EXIT, "next")
+                .addTransition("goto", OBSTACLE_EXIT, "avoid2")
+                .addTransition("goto", UNREACHABLE_EXIT, "findPath")
+
+                .addTransition("avoid2", COMPLETED_EXIT, "safe2")
+
+                .addTransition("safe2", COMPLETED_EXIT, "gotoSafe2",
+                        assign(GotoStatus.TARGET_KEY, NearestSafeStatus.TARGET_KEY))
+                .addTransition("gotoSafe2", COMPLETED_EXIT, "scanForPath")
+                .addTransition("gotoSafe2", OBSTACLE_EXIT, "avoid2")
+                .addTransition("gotoSafe2", UNREACHABLE_EXIT, "safe2")
+
+                .addTransition("scanForPath", COMPLETED_EXIT, "findPath")
+                .addTransition("scanForPath", OBSTACLE_EXIT, "avoid2")
+
+                .setParams(TIMEOUT_KEY, 2000);
+
+        JsonNode rangeNode = config.path("maxDistance");
+        StateMachineBuilder builder2 = rangeNode.isMissingNode()
+                ? builder1
+                : builder1.setParams(RandomTargetStatus.MAX_DISTANCE_KEY, rangeNode.asDouble());
+        StateMachineBuilder builder3 = point(config.path("center"))
+                .map(center -> builder2.setParams(RandomTargetStatus.CENTER_KEY, center))
+                .orElse(builder2);
+        return builder3.build("initial");
+    }
+
     static InferenceEngine sequence(JsonNode config) {
-        JsonNode targetsNode = config.path("targets");
-        if (targetsNode.isMissingNode()) {
-            throw new IllegalArgumentException("Missing targets");
-        }
-        if (!targetsNode.isArray()) {
-            throw new IllegalArgumentException("targets must be an array");
-        }
-        List<Point2D> targets = stream(targetsNode.elements()).map(node -> {
-            JsonNode xNode = node.path("x");
-            JsonNode yNode = node.path("y");
-            if (xNode.isMissingNode()) {
-                throw new IllegalArgumentException("Missing x");
-            }
-            if (yNode.isMissingNode()) {
-                throw new IllegalArgumentException("Missing y");
-            }
-            return new Point2D.Double(xNode.asDouble(), yNode.asDouble());
-        }).collect(Collectors.toList());
+        points().apply(Locator.root().path("targets")).accept(config);
+        List<Point2D> targets = points(config.path("targets"));
         return StateMachineBuilder.create()
                 .addState("nextTarget", NextSequenceStatus.create())
                 .addState("goto", GotoStatus.create())
@@ -172,7 +206,7 @@ public interface Builders {
                 .addTransition("scan", COMPLETED_EXIT, "nextTarget")
                 .addTransition("nextTarget", NextSequenceStatus.TARGET_SELECTED_EXIT, "goto",
                         assign(GotoStatus.TARGET_KEY, NextSequenceStatus.TARGET_KEY))
-                .addTransition("goto", GotoStatus.TARGET_REACHED_EXIT, "nextTarget")
+                .addTransition("goto", COMPLETED_EXIT, "nextTarget")
                 .addTransition("goto", OBSTACLE_EXIT, "scan")
                 .setParams(NextSequenceStatus.LIST_KEY, targets)
                 .build("goto");