+++ src/main/java/org/mmarini/wheelly/swing/Yaml.java
@@ -4,15 +4,20 @@ import com.fasterxml.jackson.databind.JsonNode;
 import org.mmarini.wheelly.model.ConfigParameters;
 import org.mmarini.yaml.schema.Validator;
 
+import java.awt.geom.Point2D;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static java.util.Objects.requireNonNull;
+import static org.mmarini.Utils.stream;
 
 /**
  *
  */
 public class Yaml {
+
     /**
      *
      */
@@ -63,5 +68,45 @@ public class Yaml {
     public static String engine(JsonNode root) {
         requireNonNull(root);
         return root.path("engine").asText();
+    }
+
+    /**
+     *
+     */
+    public static Validator point() {
+        return Validator.array(
+                Validator.arrayItems(Validator.number()),
+                Validator.minItems(2),
+                Validator.maxItems(2)
+        );
+    }
+
+    public static Optional<Point2D> point(JsonNode root) {
+        requireNonNull(root);
+        return root.isMissingNode() ? Optional.empty()
+                : Optional.of(new Point2D.Double(
+                root.path("0").asDouble(),
+                root.path("1").asDouble()));
+    }
+
+    public static List<Point2D> points(JsonNode root) {
+        requireNonNull(root);
+        return stream(root.elements()).flatMap(node -> point(node).stream())
+                .collect(Collectors.toList());
+    }
+
+    /**
+     *
+     */
+    public static Validator points() {
+        return Validator.arrayItems(point());
+    }
+
+    public static Validator randomPath() {
+        return Validator.objectProperties(
+                Map.of("center", point(),
+                        "maxDistance", Validator.positiveNumber()
+                )
+        );
     }
 }